add check src and dst before assigning it as the paths for work

+++ sortMyRecentPhotos.py
@@ -2,10 +2,24 @@
 
 import copy, os, shutil, sys, collections, re
 
-logFile = open('D:\\PythonPhoto\\sortedPhotos\\logFile.txt', 'w')
-logFile.write('Program started. Log file created\n\n')
+if os.path.exists(os.path.join('D:/', 'PythonPhoto', 'sortedPhotos')):
+	#mind the syntax: it is 'D:/', neither 'd', nor 'D:', nor "D:/"
+	logFile = open('D:\\PythonPhoto\\sortedPhotos\\logFile.txt', 'w')
+	logFile.write('Program started. Log file created\n\n')
+else:
+	print(os.path.join('D:', 'PythonPhoto', 'sortedPhotos') + ' doesn\'t exist')
+	logFile.write(os.path.join('D:', 'PythonPhoto', 'sortedPhotos') 
+		+ ' doesn\'t exist')
+	sys.exit()
 
-unsortedPhotos = ('D:\\PythonPhoto\\unsortedPhotos')
+if os.path.join('D:/', 'PythonPhoto', 'unsortedPhotos'):
+	unsortedPhotos = ('D:\\PythonPhoto\\unsortedPhotos')
+else:
+	print(os.path.join('D:', 'PythonPhoto', 'sortedPhotos') + ' doesn\'t exist')
+	logFile.write(os.path.join('D:', 'PythonPhoto', 'sortedPhotos') 
+		+ ' doesn\'t exist')
+	sys.exit()
+
 sortedPhotos = ('D:\\PythonPhoto\\sortedPhotos')
 logFile.write('Path to main folders created\n\n')
